refactor: move some functions to common files
